Add ClientCredentials class with the confidential client credentials type exports and exports ClientAssertion from MSAL Common
